Fix a crash in site upgrade log output due to usage of six.
Due to a bad merge, we had code for generating our evolution failure log
that was still using `six`, which wasn't imported.

This fixes up the call sites to remove that usage.

There are also a few fixes for the error output, which had extra
newlines.

Testing Done:
Simulated a failure. Verified the log was generated.

Reviewed at https://reviews.reviewboard.org/r/12773/