Add a new JS bundle for UI elements.
As we move things around in the JavaScript codebase, we'd like a place
to put any generic UI widgets. We already have a `ui/` directory in the
tree, and this change adds a new bundle for us to use for it.

Testing Done:
Loaded a page and saw that `/static/rb/js/ui/index.js` was included in
the `<script>` tags.

Reviewed at https://reviews.reviewboard.org/r/12807/