Fix a failure in js-tests.
The tests for the comment dialog were failing in Firefox due to a missed
spy. The problem is that the text editor was calling `.focus()` on the
jQuery-wrapped element when using plain text editors, but the spy was
looking for the element itself.

This change updates the `TextAreaWrapper` to call `.el.focus()` instead
of `.$el.focus()`. This makes things consistent between plain text areas
and codemirror versions, so the spy works in both cases.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/12826/