Reorganize settings for the diff viewer.
The diff viewer settings page is one of the oldest pages in the admin
UI, and its organization left much to be desired. Some upcoming work
will be adding to this page, and as such, it's time to clean it up.

This now groups the settings into 3 fieldsets:

* Appearance
* Limits
* Advanced

Appearance contains the checkboxes for trailing whitespace and syntax
highlighting, plus the field for mapping file extensions to
highlighters.

Limits contains the diff size limit and the max lines for syntax
highlighting.

Advanced contains everything it had before, plus the "Show all
whitespace for" field, which should generally no longer be used.

Some of the field labels and help text have been tweaked to be more
clear and to provide recommendations.

Testing Done:
Verified the page's contents on desktop and mobile.

Checked for spelling errors.

Reviewed at https://reviews.reviewboard.org/r/12781/