Update action extension hooks and unit tests.
This is the final change for the new actions framework, covering the
extension hooks and unit tests.

I've repurposed the old `ActionHook` (which was just used as a generic
base class) to be the new entry point for actions. This simply takes a
list of `BaseAction` instances and handles registering and unregistering
as appropriate.

The old legacy action hooks have been updated to translate the
dictionary definitions into BaseAction instances, which are then passed
upward to the ActionHook base class. These are all now treated as
deprecated, and will warn appropriately.

This also fixes up a variety of issues discovered with the unit tests.
The majority of these fixes are in the extension tests, but this also
involved updating the review request actions to use `apply_to` to limit
their rendering so that we wouldn't attempt to render their JavaScript
side on pages that were not review-request related.

Testing Done:
- Built an extension that used all of the hooks, both the modern and
  legacy. Saw that everything worked appropriately.
- Ran python unit tests.
- Ran js-tests.
- Built codebase and manual documentation and checked all related
  pages for content and spelling.

Reviewed at https://reviews.reviewboard.org/r/12775/