Add url_name parameter to TestCase.create_http_request
In order to be able to use `create_http_request` for tests related to
URL names (for example, extensions or actions that use `apply_to`), we
need to set the URL name on the `ResolverMatch`. This adds a parameter
to the test helper to let us set that.

This also fixes up usage with local sites, where we were setting
`Request.local_site` but not `Request._local_site_name`.

Testing Done:
Used in conjunction with other changes.

More testcase

Reviewed at https://reviews.reviewboard.org/r/12823/