Split up extensions/hooks.py
`hooks.py` was getting very large and unwieldy. This change splits it up
into a bunch of individual files, with an index for forwarding
imports.

This does not make any changes to the hook implementations except for
adding a couple `is not None` assertions to satisfy static type
checkers. I'm planning on following this up later with a change to add
typing to these.

Testing Done:
- Ran unit tests.
- Built codebase docs and checked output.

Reviewed at https://reviews.reviewboard.org/r/12837/