Fix post-commit UI for small Mercurial repositories.
The way that the initial commit was represented when fetching lists of
commits for Mercurial repositories wasn't correct, reporting an ID of 40
consecutive zeros, instead of an empty parent ID. The result of this is
that when the "New Review Request" UI hit the end of the commit list, it
would end up in an infinite loop requesting new commits starting at
`00000...`.

This change fixes all affected backends to convert the 000 commit to a
proper empty string.

Testing Done:
- Loaded the /r/new/ UI for a tiny Mercurial repository and saw that it
  no longer made infinite requests for commits.
- Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/12843/