Fix a regression in confirming trust when configuring repositories.
Review Board 4 introduced a regression when confirming trust in
repository certificates or SSH keys. While this isn't used particularly
often (as most people now interface with services like GitHub or with
repositories that don't require this kind of verification), it has been
noticed recently by a couple different parties.

The problem was that, when clicking "I trust this host," blank form data
was being sent along. This happened because the fields on the form are
disabled when this prompt is shown, and disabled fields aren't sent in
form data.

We had code meant to deal with this. It re-enabled the disabled fields
when confirming. However, a restyle removed the old class name on the
alert, preventing this from working.

This change simply adds back the old class name, allowing this to work
again. We're not worrying about modern naming at this stage, since the
whole form is being rewritten for Review Board 6.

Testing Done:
Triggered a host verification prompt. Verified that the form fields
were no longer disabled and the form data was properly sent along.

Reviewed at https://reviews.reviewboard.org/r/12829/