Improve the intro to the user docs, and improve some links.
This expands the introductory docs for the User Guide for Review Board,
providing a high-level summary of Review Board and some notable
integrations and features, before proceeding to linking to other
sections of the documentation.

This can help teach people more about the benefits of the product
without jumping through all the pages first, or backtracking to other
parts of the site.

Some of the links have been renamed as well, helping keep the links on
this page in sentence casing. One link, "Reviewing diffs," has been
changed to "Reviewing code (diffs)", to help clarify that page.

Testing Done:
Built and viewed the docs. Checked for bad links, spelling errors, and
build errors.

Reviewed at https://reviews.reviewboard.org/r/12839/