Rework the Review docs to better define the structure.
This updates the introductory docs on reviews to better show the
up-front structure of a review (Ship It, header, comments, footer),
numbering them, and then ordering the sections in the same way.

The contents have been rewritten a bit to flow a bit better and to link
to relevant documentation.

Testing Done:
Built and viewed the docs. Checked for spelling errors, bad links, and
build errors.

Reviewed at https://reviews.reviewboard.org/r/12840/