Rewrite the docs for creating site directories.
This is a full rewrite of the documentation for creating sites. It
provides far better coverage for all steps.

The database setup now mentions MariaDB, recommends `utf8mb4` for MySQL,
and provides more concise installation steps for Postgres.

A new section on SELinux was added, providing a full walkthrough on
properly enabling SELinux policies for HTTP daemons and
registering/applying the right permissions on site directories.

Another new section on crontabs was added, providing setup steps and an
example of the configuration file.

The largest changes have to do with the web server configuration.

We now provide a full guide on the following web servers:

* Apache + mod_wsgi
* Nginx + Gunicorn
* Nginx + uWSGI

Each of these go into the setup steps, point to the configuration files
generated with Review Board 5.0.3+, and embeds sample configuration
files.

This should do a lot to help people get set up on a modern deployment,
especially if they're aiming not to use Apache.

Testing Done:
Built the docs and read through them.

Checked for bad links, spelling errors, and build errors.

Tested the configurations locally with Apache, Nginx, Gunicorn, and
uWSGI.

Reviewed at https://reviews.reviewboard.org/r/12846/