Avoid triggering Pipeline when running unit tests.
When rendering full pages in unit tests, we were triggering Pipeline's
media compilation. This wasn't a problem before, because we had overridden
the list of Pipeline compilers, but we now hard-code Rollup if we opt in.
Rollup ends up triggering some long compilations in this environment, and
potentially was triggering some bad behavior somewhere, as it was resulting
in some stalled tests that had to be force-killed.

We now skip Pipeline entirely when running unit tests.

Testing Done:
All unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/12862/