Fix a unit test regression from the NIS change.
The NIS change removed the `nis` import in the unit tests, in favor of
accessing the `nis` attribute on the backend. A particular code path was
updated but not included in the commit.

This change adds that missing update. It also removes one more usage of
an `import`, and fixes some type checker issues in the unit tests.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/12871/