Fix a security issue with LDAP authentication with anonymous binds.
An issue was found that enables a user to log in as another user when
LDAP is configured and all the following criteria are met:

1. LDAP is enabled for authentication.

2. The LDAP server supports anonymous binds (allowing completely
   unauthenticated access to the LDAP directory)

3. The LDAP server considers a bind attempt without a DN but with a
   password as always being successful. In our testing, this is
   uncommon, and likely a security issue in an LDAP server
   implementation, but is key to the vulnerability.

4. The target user exists in the local Review Board database but has no
   entry in LDAP.

5. The server is running python-ldap 3.0 or higher (which is required in
   Review Board 4 and up, but not in Review Board 3, which is receiving
   this same fix).

This does not apply to Active Directory.

The cause of this issue was a missing check for a computed DN before
authenticating the user. If the DN lookup for a user failed, a `None`
value would be passed to a bind request, along with the password. If
successful, we'd return the matching user.

In local testing with anonymous binds enabled, a combination of an empty
DN and a non-empty password would result in an Invalid Credentials
error, but on one user's server, it did not. This prevented an error
handler from being invoked.

We believe the attack surface here is pretty minimal for the following
reasons:

1. Anonymous binds are a security risk to begin with, and many companies
   should have LDAP secured or they risk other attack vectors and PII
   leaks.

2. Most production Review Board servers using LDAP only have a single
   local-only user (the default `admin`), which matches the default
   admin username in many LDAP installs.

3. The server would need to accept a NULL DN with a password as being a
   valid bind request, and that doesn't seem to be the cause in the
   latest versions of any of the most commonly-used LDAP servers.

However, this could be a problem in some installs, and we recommend
everybody using LDAP upgrade to the latest 3.x, 4.x, or 5.x releases
immediately.

Testing Done:
Tested with the latest versions of SLAPD/OpenLDAP, the most
commonly-used LDAP servers. Waiting on details from a user about
the software/version they used.

Couldn't reproduce the original problem, but I did simulate the
conditions via unit tests, and verified the fix.

Reviewed at https://reviews.reviewboard.org/r/12855/