Hide change description field in unified banner depending on mode.
It was a little bit confusing to have the change description field
displayed when the selected draft mode was unrelated to review request
changes. This updates the visibility calculations so we hide that field
when the banner is controlling only review or reply drafts.

Testing Done:
Flipped around with different modes and saw the change description field
properly visible only when the selected mode includes the review request
changes.

Reviewed at https://reviews.reviewboard.org/r/12876/