Scroll to related review when switching banner mode to a reply.
The new unified banner makes it much easier to not forget about draft
replies (which could previously have been hidden deep within a very long
review request page full of reviews and updates), but just seeing that
the reply exists still makes it a little annoying to find it.

This change makes it so switching modes in the banner to a review reply
will scroll the page to the review being replied to.

Testing Done:
Set up a long review request with lots of updates and reviews, and
created some draft replies. Saw that switching modes correctly scrolled
the page (and looked good, due to the scroll-behavior setting).

Reviewed at https://reviews.reviewboard.org/r/12900/