Make wording consistent for review request close terminology.
As part of the actions rework, I changed "Close -> Submitted" to
"Close -> Completed". The word submitted is pretty unique to Perforce,
and has been a source of confusion for users over the years, especially
since in some contexts we used the word "submitted" to mean "put up for
review".

This change fixes up other uses of the "Submitted" term, including in
the dashboard and the manual.

We still use "submitted" on the backend, including in webhook
notifications. This won't be changing because it's API.

Testing Done:
- Used the dashboard bulk mode and saw the correct wording.
- Built the manual.

Reviewed at https://reviews.reviewboard.org/r/12879/