Fix hover for edit icons in the main review request area.
We've had a long standing style bug where the edit icons on the summary
and description fields would show an underline under the field-required
asterisk when they were hovered over. This wasn't a problem for the
owner field because we have a special rule for the
review-request-details table that prevents hover underlines for all
links.

This change adds a rule for edit icons to prevent underlines on hover.

Testing Done:
Hovered over the edit icon for the summary and description fields and no
longer saw blue underlines under the red asterisks.

Reviewed at https://reviews.reviewboard.org/r/12881/