Plumb through redirect URL when performing SAML autologin.
When an SSO backend is configured to log in automatically, the redirect
URL was being lost. We already have redirect URLs properly used in the
SAML login view and SAML handshake, so this just required passing it
along in the redirect from the login view to the SAML login view.

Testing Done:
Configured autologin for SAML and accessed /account/preferences/ while
logged out. Saw that after the SSO flow, I was properly redirected to
the account settings page rather than the dashboard.