Fix our approach to mixing in classes in BaseResource and ReviewReply.
`RB.BaseResource` and `RB.ReviewReply` both mix in other classes, but
were doing so after the call to `.extend()`. This meant that the mixin
wasn't part of the original state going into the prototype. While not a
problem in Backbone, it does impact Spina, as the wrong prototype ends
up being used.

This change fixes up these two classes to apply mixins the same way that
other classes do, inline in a call to `.extend()`.

Testing Done:
Unit tests pass.

Tested this with the in-progress Spina 2 work, which supports `.extend()`.

Reviewed at https://reviews.reviewboard.org/r/12902/