Fix the wording and links on the Trojan Source alert.
The Trojan Source alert in the diff viewer mistakenly alluded to all
threats being due to invisible Unicode characters. Originally this
wording was intended for the zero-width spaces and bi-directional
characters.

On top of this, due to an error and a bad copy/paste, the CVE links were
all broken.

This change updates the wording to be more general, and fixes all CVE
links.

Testing Done:
Tested each of the links.

Reviewed at https://reviews.reviewboard.org/r/12886/