Improve error text when elasticsearch is unavailable.
When validating the Elasticsearch configuration, we'd check if the
`elasticsearch` module was available and was of a compatible version. If
either failed, we'd display a generic error about ensuring you have a
compatible `elasticsearch`.

This change updates validation to show an error specific to each case.
Both include the list of packages to install for Elasticsearch support.

Unit tests have been added to ensure validation works and to track any
changes to our list of compatible versions.

Testing Done:
Manually tested this without `elasticsearch` installed, and with an
incompatible version.

Unit tests passed.

Reviewed at https://reviews.reviewboard.org/r/12888/