Add a missing configuration for Apache + mod_wsgi.
Review Board 5.0.3 introduced new configuration files for web servers,
but somehow, the final version of my commit for this was missing the
changes to `apache-wsgi.conf.in`. This was a real problem, since
`rb-site` had changes that had to be paired with the new configuration.
The result was that default Apache installs likely didn't work for most
people.

The missing change to this file was found in `git reflog`. This change
re-introduces the new version of the file.

Testing Done:
Tested this configuration previously when building the original change.

Reviewed at https://reviews.reviewboard.org/r/12917/